Update SystemClock_NTP example
